% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bestbooksever_df}
\alias{bestbooksever_df}
\title{Metdata for the first 100 good reads books from the best books ever webpage}
\format{
\describe{
\item{Rank}{Rank given to a book.}
\item{Book Names}{Name of book.}
\item{Series}{If a book is part of a series, it will list the serie name. If else, it will be NA.}
\item{Author}{Name of Author.}
\item{Average Rating}{Average rating on the scale of 1 to 5, with 5 being the best rating.}
\item{Total Rating}{Total number of ratings for a book.}
\item{Cover ID}{Book code needed to get the cover image.}
}
}
\source{
\url{https://www.goodreads.com/list/show/1.Best_Books_Ever}
}
\usage{
bestbooksever_df
}
\description{
Metdata for the first 100 good reads books from the best books ever webpage
}
\keyword{datasets}
